ref(tests): Add GitHubWebhookCodeReviewTestCase helper class
- Created GitHubWebhookCodeReviewTestCase base class in testutils/helpers/github.py
- Added code_review_setup() context manager for test setup
- Added _send_webhook_event() helper method to eliminate code duplication
- Updated test_check_run.py and test_issue_comment.py to use the new base class
- All tests passing with proper type hints
