+++ src/sentry/testutils/helpers/github.py
@@ -10,6 +10,7 @@ from datetime import datetime, timedelta
 from typing import Any
 from uuid import uuid4
 
+import orjson
 from django.http.response import HttpResponseBase
 
 from sentry import options
@@ -140,12 +141,35 @@ class GitHubWebhookCodeReviewTestCase(GitHubWebhookTestCase):
 
     @contextmanager
     def code_review_setup(
-        self, features: Collection[str] | Mapping[str, Any] | None = None
+        self,
+        features: Collection[str] | Mapping[str, Any] | None = None,
+        forward_to_overwatch: bool = False,
     ) -> Generator[None]:
         """Helper to set up code review test context."""
         self.organization.update_option("sentry:enable_pr_review_test_generation", True)
+        features_to_enable = self.CODE_REVIEW_FEATURES if features is None else features
         with (
-            self.feature(features or self.CODE_REVIEW_FEATURES),
-            self.options({"github.webhook.issue-comment": False}),
+            self.feature(features_to_enable),
+            self.options({"github.webhook.issue-comment": forward_to_overwatch}),
         ):
             yield
+
+    def _send_webhook_event(
+        self, github_event: GithubWebhookType, event_data: bytes | str
+    ) -> HttpResponseBase:
+        """Helper to send a GitHub webhook event."""
+        self.event_dict = (
+            orjson.loads(event_data) if isinstance(event_data, (bytes, str)) else event_data
+        )
+        repo_id = int(self.event_dict["repository"]["id"])
+
+        integration = self.create_github_integration()
+        self.create_repo(
+            project=self.project,
+            provider="integrations:github",
+            external_id=repo_id,
+            integration_id=integration.id,
+        )
+        response = self.send_github_webhook_event(github_event, event_data)
+        assert response.status_code == 204
+        return response
