+++ tests/sentry/seer/code_review/test_webhooks.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timedelta, timezone
-from typing import Any
 from unittest.mock import MagicMock, patch
 
 import pytest
