+++ tests/sentry/seer/code_review/webhooks/test_check_run.py
@@ -1,7 +1,6 @@
 from unittest.mock import MagicMock, patch
 
 import orjson
-from django.http.response import HttpResponseBase
 
 from fixtures.github import (
     CHECK_RUN_COMPLETED_EVENT_EXAMPLE,
@@ -14,26 +13,6 @@ from sentry.testutils.helpers.github import GitHubWebhookCodeReviewTestCase
 
 class CheckRunEventWebhookTest(GitHubWebhookCodeReviewTestCase):
     """Integration tests for GitHub check_run webhook events."""
-
-    def _send_webhook_event(
-        self, github_event: GithubWebhookType, event_data: bytes | str
-    ) -> HttpResponseBase:
-        """Helper to send a GitHub webhook event."""
-        self.event_dict = (
-            orjson.loads(event_data) if isinstance(event_data, (bytes, str)) else event_data
-        )
-        repo_id = int(self.event_dict["repository"]["id"])
-
-        integration = self.create_github_integration()
-        self.create_repo(
-            project=self.project,
-            provider="integrations:github",
-            external_id=repo_id,
-            integration_id=integration.id,
-        )
-        response = self.send_github_webhook_event(github_event, event_data)
-        assert response.status_code == 204
-        return response
 
     @patch("sentry.seer.code_review.webhooks.task.process_github_webhook_event")
     def test_base_case(self, mock_task: MagicMock) -> None:
